Extract a new `domApiCompatExtra` module
