+++ DevToolkt/domApiCompatExtra/build.gradle.kts
@@ -1,0 +1,16 @@
+plugins {
+    alias(libs.plugins.kotlin.multiplatform)
+}
+
+group = "dev.toolkt"
+
+repositories {
+    mavenCentral()
+}
+
+kotlin {
+    js(IR) {
+        browser {
+        }
+    }
+}
