+++ DevToolkt/domApiCompatExtra/src/jsMain/kotlin/org/w3c/dom/extra/utils.kt
@@ -1,0 +1,4 @@
+package org.w3c.dom.extra
+
+@Suppress("NOTHING_TO_INLINE")
+inline fun jsObject(): dynamic = js("({})")
