+++ app/src/jsMain/resources/index.html
@@ -9,9 +9,9 @@
 <script src="https://unpkg.com/path-data-polyfill@1.0.10/path-data-polyfill.js"></script>
 <script src="kotlin-kotlin-stdlib.js"></script>
 <script src="kotlin_org_jetbrains_kotlin_kotlin_dom_api_compat.js"></script>
-<script src="kotlinx-benchmark-kotlinx-benchmark-runtime.js"></script>
 <script src="kotlinx-serialization-kotlinx-serialization-core.js"></script>
 <script src="kotlinx-serialization-kotlinx-serialization-json.js"></script>
+<script src="DevToolkt-domApiCompatExtra.js"></script>
 <script src="DevToolkt-core.js"></script>
 <script src="DevToolkt-math.js"></script>
 <script src="DevToolkt-geometry.js"></script>
